requirements.txt
beautifulsoup4==4.12.3
jieba==0.42.1
numpy==2.2.1

Pillow==11.0.0
pyecharts==2.0.7
Requests==2.32.3
streamlit==1.41.1
streamlit_echarts==0.4.0
